requirements.txt
Flask==3.0.0
suntime==1.2.5
requests==2.31.0
pyswisseph==2.10.3.2
pytz==2023.3
pytest==7.4.4
